+++ R/01-workflowstep-class.R
@@ -5,37 +5,39 @@
 #' Create a new workflow step object
 #'
 #' This object defines a single step within a workflow,
-#' including its unique identifier, name, operation, parameters, and other metadata.
-#' @param id             An integer identifier for the step.
-#' @param operation      A character string specifying the name of the function to execute for
+#' including its unique identifier, name, command, parameters, and other metadata.
+#' @param id             An integer identifier for the step. MUST BE THE NUMBER OF THE STEP IN THE WORKFLOW
+#' @param command        A character string specifying the name of the function to execute for
 #'                       this step, e.g. "strsplit". This function must exist in the loaded
 #'                       name space or in a custom script environment.
 #' @param name           A human-readable name for the step. Defaults to "Step <id>".
 #' @param label          A label for the step, used in UIs. Defaults to the same as `name`.
 #' @param comments       A character string with comments or description for the step.
-#' @param params         A list of parameters to pass to the operation function.
+#' @param args           The original argument string from the workflow file, for reference.
+#' @param params         A list of parameters to pass to the command function.
 #' @param loop           A character string indicating if the step should be looped over.
 #'                       Can be "yes", "no", or "auto".
-#' @param env            An environment to look up the operation function. Default is the parent
+#' @param env            An environment to look up the command function. Default is the parent
 #'                       frame.
 #' @param ...            Additional metadata to store with the step.
 #' @return A `workflowstep` object.
 #' @export
 new_workflowstep <- function(
   id,
-  operation,                     # function name (incl. custom name in script), exist in name space
+  command,                       # function name (incl. custom name in script), exist in name space
   name            = NULL,
   label           = NULL,
   comments        = "",
+  args            = "",          # original argument string from workflow file, for reference
   params          = list(),      # free-form list for step-specific parameters
   loop            = "",          # loop variable name (if any)
-  env             = parent.frame(),  # where to look up operation
+  env             = parent.frame(),  # where to look up command
   ...
 ) {
   if (is.null(name)) name <- paste("Step", id)
   if (is.null(label)) label <- name
 
-  resolve_operation(operation, env = env)
+  resolve_operation(command, env = env)
 
   structure(
     list(
@@ -43,7 +45,8 @@ new_workflowstep <- function(
       name            = name,
       label           = label,
       comments        = comments,
-      operation       = operation,
+      command         = command,
+      args            = args,
       params          = params,
       loop            = loop,
       env             = env,
@@ -70,7 +73,7 @@ print.workflowstep <- function(x, ...) {
   if (nzchar(x$comments)) {
     cat("  comments:       ", x$comments, "\n", sep = "")
   }
-  cat("  operation:      ", x$operation, "\n", sep = "")
+  cat("  command:      ", x$command, "\n", sep = "")
   if (length(x$params)) {
     cat("  params:\n")
     utils::str(x$params, indent.str = "    ")
@@ -108,8 +111,8 @@ as.data.frame.workflowstep <- function(x, ...) {
     "Name"          = x$name,
     "Label"         = x$label,
     "Comments"      = x$comments,
-    "Function"      = x$operation,
-    "Parameters"    = flatten_params(x$params),
+    "Function"      = x$command,
+    "Parameters"    = x$args, # flatten_params(x$params),
     stringsAsFactors = FALSE
   )
 }
@@ -120,8 +123,8 @@ map_field <- function() {
     Name = "name",
     Label = "label",
     Comments = "comments",
-    Function = "operation",
-    Parameters = "params"
+    Function = "command",
+    Parameters = "args"
   )
 }
 
@@ -129,6 +132,7 @@ map_field <- function() {
 #'
 #' @param x A `workflowstep` object.
 #' @param field The name of the field to retrieve (e.g., "name", "comments").
+#' @param with_map_field Logical, whether to map the field name using `map_field()`. Defaults to `TRUE`.
 #' @param ... Additional arguments (not used).
 #' @return The value of the specified field from the workflowstep.
 #' @export
@@ -140,14 +144,15 @@ get_field.workflowstep <- function(x, field, with_map_field = TRUE, ...) {
 }
 
 #' Update a workflow step
-#' 
+#'
 #' This method allows updating specific fields of a `workflowstep` object, such as its name,
-#' label, comments, operation, parameters, or loop configuration.
-#' 
+#' label, comments, command, parameters, or loop configuration.
+#'
 #' @param x The `workflowstep` object to update.
+#' @param workflow_file_paths The paths to the workflow files, used for updating related files.
 #' @param value The new value to assign to the specified field.
 #' @param entry The name of the field to update. Must be one of "name", "label", "comments",
-#'  "operation", "params", or "loop".
+#'  "command", "params", or "loop".
 #' @param with_map_field Logical, whether to map the field name using `map_field()`.
 #'  Defaults to `TRUE`.
 #' @param ... Additional arguments (not used).
@@ -155,6 +160,7 @@ get_field.workflowstep <- function(x, field, with_map_field = TRUE, ...) {
 #' @export
 update.workflowstep <- function(
   x,
+  workflow_file_paths,
   value,
   entry,
   with_map_field = TRUE,
@@ -182,28 +188,53 @@ update.workflowstep <- function(
   }
 
   # validate value
-  if (entry %in% c("name", "label", "comments", "operation")) {
-    if (!is.character(value) || length(value) != 1L) {
-      stop(sprintf("'%s' must be a single character string.", entry), call. = FALSE)
-    }
-  } else if (entry == "params") {
-    #browser()
-    if (!is.character(value) || !is.list(value)) {
-      #stop("'params' can only be a list o.", call. = FALSE)
-    }
-    # we need to extract param from character string
+  if (!is.character(value) || length(value) != 1L) {
+    stop(sprintf("'%s' must be a single character string.", entry), call. = FALSE)
   }
+
+  # update the specified field in the workflowstep object
   x[[entry]] <- value
+
+  # update related objects
+  if (entry == "args") {
+    # update params also here, not only at runtime, to keep them in sync
+    # we should update files also here, not only at runtime, to keep them in sync
+    x$params <- extract_params_from_arg_string(
+      args_string = value,
+      loop = x$loop,
+      step_i = x$id,
+      wf_file_paths = workflow_file_paths # used to extract concrete values for inputs
+    )
+  }
+
+  # update commands.json file
+  # get i-th entry from commands file, update it and write back to file
+  commands_list <- read_json_if_exists(path = workflow_file_paths$commands_path)
+
+  # update i-th step
+  logDebug(
+    "Updating workflow file '%s': step %d, entry '%s' with value '%s'",
+    basename(workflow_file_paths$commands_path),
+    x$id,
+    entry,
+    value
+  )
+  # do all entries match?
+  # align id <> entry? Can we update the id right now? -> no, cannot yet change order or id/entry
+  commands_list[[x$id]][[entry]] <- value
+  write_json(commands_list, path = workflow_file_paths$commands_path, auto_unbox = TRUE)
+
+  # return updated workflowstep
   x
 }
 
 resolve_operation <- function(op_name, env) {
-  PEITHO:::logDebug("  Resolving operation function: %s", op_name)
+  PEITHO:::logDebug("  Resolving command function: %s", op_name)
   if (!is.character(op_name) || length(op_name) != 1L || !nzchar(op_name)) {
-    stop("'operation' must be a non-empty character string.", call. = FALSE)
+    stop("'command' must be a non-empty character string.", call. = FALSE)
   }
   if (!exists(op_name, mode = "function", envir = env, inherits = TRUE)) {
-    stop("Operation '", op_name, "' not found in given environment.", call. = FALSE)
+    stop("Command '", op_name, "' not found in given environment.", call. = FALSE)
   }
   get(op_name, envir = env, mode = "function", inherits = TRUE)
 }
@@ -221,7 +252,7 @@ run_with_error <- function(fn, args) {
 #' with the result or error from the step execution.
 #' @param x  A `workflowstep` object representing the step to execute.
 #' @param state A `workflowstate` object representing the current state of the workflow.
-#' @param env   An environment to look up the operation function. Defaults to the step's
+#' @param env   An environment to look up the command function. Defaults to the step's
 #'  own env or the caller's env.
 #' @param ...   Additional arguments (not used).
 #' @return A `workflowsteprun` object recording the execution of the step.
@@ -229,7 +260,7 @@ run_with_error <- function(fn, args) {
 run.workflowstep <- function(
   x,
   state,
-  env = NULL,  # where to look up operation
+  env = NULL,  # where to look up function
   ...
 ) {
   if (!inherits(state, "workflowstate")) {
@@ -241,7 +272,7 @@ run.workflowstep <- function(
   }
   # 1) resolve the function
   # for a package you might use: env = asNamespace("PEITHO")
-  fn <- resolve_operation(x$operation, env)
+  fn <- resolve_operation(x$command, env)
 
   # 2) assemble arguments
   params <- x$params
@@ -249,12 +280,13 @@ run.workflowstep <- function(
 
   args <- list()
 
-  PEITHO:::logDebug("  Extract arguments for operation '%s'", x$operation)
+  PEITHO:::logDebug("  Extract arguments for command '%s'", x$command)
   for (param in params) {
     if (!inherits(param, "operationparam")) {
       stop("All entries in 'params' must be of class 'operationparam'.", call. = FALSE)
     }
 
+    # unpack params object into args
     arg_list <- extract_arg_list(param, state = state)
     args <- c(args, arg_list)
   }
@@ -280,8 +312,8 @@ run.workflowstep <- function(
   # if loop_param and loop_arg disagree, throw error
   if (!identical(loop_param_indices, arg_list_indices)) {
     PEITHO:::logWarn(
-      "WARNING! Detected list argument(s) for operation '%s', but 'loop' is set to '%s'.",
-      x$operation,
+      "WARNING! Detected list argument(s) for command '%s', but 'loop' is set to '%s'.",
+      x$command,
       params[[arg_list_indices[1]]]$loop
     )
   }
@@ -289,7 +321,7 @@ run.workflowstep <- function(
   # 3) actually call the function, if needed then in a loop
   if (any(is_param_config_loop)) {
     PEITHO:::logDebug(
-      "  Running operation: WITH LOOPING over argument index %d",
+      "  Running command: WITH LOOPING over argument index %d",
       loop_param_indices
     )
     loop_index <- loop_param_indices[1]
@@ -302,7 +334,7 @@ run.workflowstep <- function(
     results <- lapply(runs, `[[`, "output")
     errors  <- lapply(runs, `[[`, "error")
 
-    PEITHO:::logInfo("  %d loop iterations for operation '%s':", length(runs), x$operation)
+    PEITHO:::logInfo("  %d loop iterations for command '%s':", length(runs), x$command)
     max_result_length <- max(lengths(results))
     if (max_result_length > 1L) {
       PEITHO:::logWarn(
@@ -320,15 +352,15 @@ run.workflowstep <- function(
       error  = errors
     )
   } else {
-    PEITHO:::logDebug("  Running operation: NO LOOPING")
+    PEITHO:::logDebug("  Running command: NO LOOPING")
 
     run <- run_with_error(fn, args) # <--- RUN FUNCTION HERE, single run
 
     # check if result has length > 1 or not
     is_single_result <- length(run$output) == 1L
     PEITHO:::logInfo(
-      "  Operation '%s': %s result%s",
-      x$operation,
+      "  Command '%s': %s result%s",
+      x$command,
       if (is_single_result) "single" else length(run$output),
       if (is_single_result) "" else "s"
     )