+++ R/02-WebText-factory.R
@@ -9,8 +9,11 @@ peitho_user_agent <- function() {
 #' Fetch and parse web text from a URL
 #'
 #' This function retrieves the HTML content from the specified URL,
-#' extracts text based on the provided CSS selector, and returns a
-#' `WebText` object containing the extracted text and metadata.
+#' parses it to extract the main text content, and returns it as a character vector.
+#' It also handles errors and warnings gracefully, providing informative messages if
+#' the request fails or if the HTML structure is not as expected.
+#' By default, it returns only the extracted text blocks, but it can also return a
+#' full `WebText` object containing metadata if `return_text_blocks_only` is set to `FALSE`.
 #'
 #' @param url A character string specifying the URL to fetch.
 #' @param timeout_sec An integer specifying the timeout for the HTTP
@@ -110,7 +113,7 @@ fetch_WebText <- function(
     }
   }
 
-  if (nzchar(text_blocks)) {
+  if (length(text_blocks) > 0 && nzchar(text_blocks)) {
     text_blocks <- stringr::str_squish(text_blocks)
   }