+++ R/02-results_table-module.R
@@ -1,11 +1,16 @@
 #' Results Table UI Module
 #'
 #' @param id Shiny module id
+#' @param title Title to display above the table
 #' @return Shiny UI output for the results table
 #' @export
-results_table_ui <- function(id) {
+results_table_ui <- function(id, title = "") {
   ns <- NS(id)
-  tableOutput(ns("results_table"))
+  tagList(
+    tags$h4(title),
+    tableOutput(ns("results_table")),
+    uiOutput(ns("select"))
+  )
 }
 #' Results Table Server Module
 #'
@@ -16,10 +21,92 @@ results_table_ui <- function(id) {
 #' @export
 results_table_server <- function(id, wf_run, is_active_tab) {
   moduleServer(id, function(input, output, session) {
+    ns <- session$ns
+
     output$results_table <- renderTable({
       wfr <- wf_run()
       if (is.null(wfr)) return(NULL)
       as.data.frame(wfr)
     }, rownames = TRUE)
+
+    output$select <- renderUI({
+      # hide the edit UI when no workflow is loaded
+      if (is.null(wf_run())) return(NULL)
+
+      field_select_ui(ns("field_select"))
+    })
+
+    sel <- field_select_server("field_select", reactive(as.data.frame(wf_run())), is_active_tab)
+
+    observe({
+      req(isTRUE(is_active_tab()), wf_run())
+      PEITHO:::logDebug("%s: Selected field: %s", id, paste(sel$selected_row(), sel$selected_column(), sep = ", "))
+    })
+  })
+}
+
+field_select_ui <- function(id) {
+  ns <- NS(id)
+  tagList(
+    tags$hr(),
+    tags$h4("Select Field"),
+    fluidRow(
+      column(
+        3,
+        selectInput(ns("row_select"), "Select row", choices = NULL, selected = NULL)
+      ),
+      column(
+        3,
+        selectInput(ns("column_select"), "Select column", choices = NULL, selected = NULL)
+      )
+    )
+  )
+}
+
+field_select_server <- function(id, df, is_active_tab) {
+  moduleServer(id, function(input, output, session) {
+    observe({
+      req(isTRUE(is_active_tab()))
+
+      if (is.null(df())) {
+        PEITHO:::logDebug("%s: Reset inputs if empty wf", id)
+        # hide the step select input when no df is loaded
+        updateSelectInput(session, "row_select", choices = NULL, selected = NULL)
+        updateSelectInput(session, "column_select", choices = NULL, selected = NULL)
+      }
+    })
+
+    observe({
+      req(isTRUE(is_active_tab()), df())
+      PEITHO:::logDebug("%s: Update inputs", id)
+
+      df_val <- df()
+      row_choices <- rownames(df_val)
+
+      if (isTRUE(input$row_select %in% row_choices)) {
+        selected_row <- input$row_select
+      } else {
+        selected_row <- NULL
+      }
+      updateSelectInput(
+        session, "row_select", choices = row_choices, selected = selected_row
+      )
+
+      column_choices <- colnames(df_val)
+
+      if (isTRUE(input$column_select %in% column_choices)) {
+        selected_column <- input$column_select
+      } else {
+        selected_column <- NULL
+      }
+      updateSelectInput(
+        session, "column_select", choices = column_choices, selected = selected_column
+      )
+    })
+
+    list(
+      selected_row = reactive(input$row_select),
+      selected_column = reactive(input$column_select)
+    )
   })
 }